requirements.txt
fastapi~=0.109
uvicorn~=0.27
pydantic~=2.0
pydantic-settings~=2.0
mcp~=1.0
httpx~=0.27
openstacksdk~=3.0
pytest~=8.0
pytest-asyncio~=0.23
